# openapi-agent-actions.yaml
...
openapi: 3.1.0
info:
  title: QuranChain Omar Autonomous Agent Actions API
  version: 1.3.0
  description: >
    Granular (single-action) endpoints for an autonomous agent persona ("QuranChain AI" + "Omar AI")
    that can run tasks, create sub-agents, manage integrations, and manage secrets safely.

    Safety model:
    - No password harvesting or "creating credentials" by guessing/inventing logins.
    - External account provisioning is only via approved API/OAuth flows and always requires approval.
    - Secrets can be stored only in a vault; values are never returned.
servers:
  - url: https://api.fullagent.com/v1
    description: Production
  - url: http://localhost:6200/v1
    description: Local Development
...
paths:
  /agents:
    post:
      operationId: createAgent
      tags: [Agent]
      summary: Create agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          description: Bad request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [Agent]
      summary: Get agent
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    patch:
      operationId: updateAgent
      tags: [Agent]
      summary: Update agent (name/description/instructions)
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAgentRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
  /agents/{agentId}/personas:
    put:
      operationId: setPersonas
      tags: [Personas]
      summary: Set personas (QuranChain AI, Omar AI)
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetPersonasRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
  /agents/{agentId}/specializations:
    put:
      operationId: setSpecializations
      tags: [Specializations]
      summary: Set specializations (QuranChain, DarCloud, Fungi Mesh, Cloudflare, Stripe)
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetSpecializationsRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
  /agents/{agentId}/autonomy:
    put:
      operationId: setAutonomy
      tags: [Autonomy]
      summary: Set autonomy policy (approvals, limits)
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetAutonomyRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
  /agents/{agentId}/messages:
    post:
      operationId: sendMessage
      tags: [Agent]
      summary: Send message (non-task)
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendMessageRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendMessageResponse" }
...
  /agents/{agentId}/tasks/{taskId}/steps:
    get:
      operationId: listTaskSteps
      tags: [Tasks]
      summary: List task steps
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TaskStep" }
...
  /agents/{agentId}/deployments:
    post:
      operationId: createDeployment
      tags: [Deployments]
      summary: Create a deployment plan/run
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDeploymentRequest" }
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deployment" }
  /agents/{agentId}/external-accounts:
    post:
      operationId: requestExternalAccount
      tags: [ExternalAccounts]
      summary: Request external account provisioning
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExternalAccountProvisionRequest" }
      responses:
        "202":
          description: Accepted (awaiting approval)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExternalAccountProvisionResult" }
...
components:
  parameters:
    agentId:
      name: agentId
      in: path
      required: true
      schema: { type: string }
...
    taskId:
      name: taskId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
    CreateAgentRequest:
      type: object
      additionalProperties: false
      required: [name, instructions]
      properties:
        name: { type: string }
        description: { type: string }
        instructions:
          type: string
          description: Include your "QuranChain AI + Omar AI" behavior here if you want it fixed at creation time.
    UpdateAgentRequest:
      type: object
      additionalProperties: false
      properties:
        name: { type: string }
        description: { type: string }
        instructions: { type: string }
    Agent:
      type: object
      additionalProperties: false
      required: [agentId, name]
      properties:
        agentId: { type: string }
        name: { type: string }
        description: { type: string }
        instructions: { type: string }
        personas:
          type: array
          items: { $ref: "#/components/schemas/Persona" }
        specializations:
          type: array
          items: { $ref: "#/components/schemas/Specialization" }
        autonomy:
          $ref: "#/components/schemas/AutonomyPolicy"
    Persona:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          enum: ["QuranChain AI", "Omar AI"]
        style: { type: string }
        purpose: { type: string }
    SetPersonasRequest:
      type: object
      additionalProperties: false
      required: [personas]
      properties:
        personas:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/Persona" }
    Specialization:
      type: string
      enum: [QuranChain, DarCloud, FungiMeshNetwork, Cloudflare, Stripe]
    SetSpecializationsRequest:
      type: object
      additionalProperties: false
      required: [specializations]
      properties:
        specializations:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/Specialization" }
    AutonomyPolicy:
      type: object
      additionalProperties: false
      properties:
        mode:
          type: string
          enum: [manual, assisted, autonomous]
        maxStepsPerTask:
          type: integer
          minimum: 1
          maximum: 500
          default: 50
        allowToolUse:
          type: boolean
          default: true
        approvalsRequiredFor:
          type: array
          items:
            type: string
            enum:
              - external_account_create
              - credential_rotate
              - stripe_money_movement
              - destructive_infra_change
              - dns_change
          default: [external_account_create, credential_rotate, stripe_money_movement]
      required: [mode]
    SetAutonomyRequest:
      type: object
      additionalProperties: false
      required: [autonomy]
      properties:
        autonomy: { $ref: "#/components/schemas/AutonomyPolicy" }
    SendMessageRequest:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message: { type: string }
        persona:
          type: string
          enum: ["QuranChain AI", "Omar AI"]
    SendMessageResponse:
      type: object
      additionalProperties: false
      required: [response]
      properties:
        response: { type: string }
        conversationId: { type: string }
...
    TaskStep:
      type: object
      additionalProperties: false
      required: [stepId, taskId, index, type, status]
      properties:
        stepId: { type: string }
        taskId: { type: string }
        index: { type: integer, minimum: 0 }
        type: { type: string }
        status:
          type: string
          enum: [pending, running, succeeded, failed, skipped]
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time }
        content: { type: string }
        metadata:
          type: object
          additionalProperties: true
...
    ApprovalRequest:
      type: object
      additionalProperties: false
      required: [approvalId, category, summary]
      properties:
        approvalId: { type: string }
        category:
          type: string
          enum:
            - external_account_create
            - credential_rotate
            - stripe_money_movement
            - destructive_infra_change
            - dns_change
        summary: { type: string }
        details:
          type: object
          additionalProperties: true
...
    CreateDeploymentRequest:
      type: object
      additionalProperties: false
      required: [title, target]
      properties:
        title: { type: string }
        target:
          type: string
          enum: [Cloudflare, Stripe, DarCloud, QuranChain, FungiMeshNetwork, Custom]
        integrationId:
          type: string
          description: Optional integration to use
        planOnly:
          type: boolean
          default: true
        desiredState:
          type: object
          additionalProperties: true
    Deployment:
      type: object
      additionalProperties: false
      required: [deploymentId, agentId, title, target, status, createdAt]
      properties:
        deploymentId: { type: string }
        agentId: { type: string }
        title: { type: string }
        target: { type: string }
        status:
          type: string
          enum: [queued, running, awaiting_approval, succeeded, failed, canceled]
        createdAt: { type: string, format: date-time }
        output: { type: string }
    ExternalAccountProvisionRequest:
      type: object
      additionalProperties: false
      required: [provider, purpose]
      properties:
        provider:
          type: string
          description: External system name (e.g., Cloudflare, Stripe, GitHub, etc.)
        purpose:
          type: string
          description: Why the account is needed
        method:
          type: string
          enum: [oauth, apiProvisioning]
          default: oauth
        connectionId:
          type: string
          description: If method=oauth, reference an OAuth connection
        requestedScopes:
          type: array
          items: { type: string }
        metadata:
          type: object
          additionalProperties: true
    ExternalAccountProvisionResult:
      type: object
      additionalProperties: false
      required: [requestId, status]
      properties:
        requestId: { type: string }
        status:
          type: string
          enum: [awaiting_approval, running, succeeded, failed]
        approval:
          $ref: "#/components/schemas/ApprovalRequest"
        accountRef:
          type: string
          description: Reference to provisioned account (no passwords)
...
    Error:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true